Migrated changes to `NetFramework`
This commit adds the latest changes also to the `NetFramework` version.

Fixed #39
