+++ source/GcodeParserSharp/GcodeParserSharp.Test/packages.config
@@ -1,14 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="CommunityToolkit.Mvvm" version="8.2.0" targetFramework="net472" />
+  <package id="CommunityToolkit.Mvvm" version="8.2.1" targetFramework="net472" />
   <package id="Cyotek.Drawing.BitmapFont" version="2.0.4" targetFramework="net472" />
   <package id="HelixToolkit" version="2.24.0" targetFramework="net472" />
   <package id="HelixToolkit.Wpf" version="2.24.0" targetFramework="net472" />
   <package id="HelixToolkit.Wpf.SharpDX" version="2.24.0" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="7.0.0" targetFramework="net472" />
-  <package id="Microsoft.Extensions.Logging.Abstractions" version="7.0.0" targetFramework="net472" />
-  <package id="MSTest.TestAdapter" version="3.0.4" targetFramework="net472" />
-  <package id="MSTest.TestFramework" version="3.0.4" targetFramework="net472" />
+  <package id="Microsoft.Extensions.Logging.Abstractions" version="7.0.1" targetFramework="net472" />
+  <package id="MSTest.TestAdapter" version="3.1.1" targetFramework="net472" />
+  <package id="MSTest.TestFramework" version="3.1.1" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net472" />
   <package id="RCoreSharp" version="1.0.9" targetFramework="net472" />
   <package id="SharpDX" version="4.2.0" targetFramework="net472" />
