--- source/GcodeParserSharp/GcodeParserSharp/Enums/SlicerName.cs
@@ -1,21 +1,0 @@
-﻿namespace AndreasReitberger.Parser.Gcode.Enums
-{
-    public enum SlicerName
-    {
-        Unkown,
-        Slic3r,
-        PrusaSlicer,
-        KISSlicer,
-        Skeinforge,
-        Cura,
-        Makerbot,
-        FlashForge,
-        Simplify3D,
-        Snapmakerjs,
-        IdeaMaker,
-        Voxelizer2,
-        CraftWare,
-        OrcaSlicer,
-        BambuStudio,
-    }
-}
