--- source/GcodeParserSharp/GcodeParserSharp/Enums/SlicerParameter.cs
@@ -1,15 +1,0 @@
-﻿namespace AndreasReitberger.Parser.Gcode.Enums
-{
-    public enum SlicerParameter
-    {
-        Volume,
-        PrintTime,
-        PrintTimeSilent,
-        PrintTimeModel,
-        FilamentUsed,
-        FilamentDiameter,
-        FilamentType,
-        FilamentDensity,
-        NozzleDiameter,
-    }
-}
