+++ source/GcodeParserSharp/GcodeParserSharp/GcodeParserSharp.csproj
@@ -10,8 +10,12 @@
 	<PackageReadmeFile>README.md</PackageReadmeFile>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="..\..\..\src\GcodeParserSharp\Enums\SlicerName.cs" Link="Enums\SlicerName.cs" />
+    <Compile Include="..\..\..\src\GcodeParserSharp\Enums\SlicerParameter.cs" Link="Enums\SlicerParameter.cs" />
     <Compile Include="..\..\..\src\GcodeParserSharp\GcodeParser.cs" Link="GcodeParser.cs" />
     <Compile Include="..\..\..\src\GcodeParserSharp\GcodeParser.Extension.cs" Link="GcodeParser.Extension.cs" />
+    <Compile Include="..\..\..\src\GcodeParserSharp\GcodeParser.Regex.cs" Link="GcodeParser.Regex.cs" />
+    <Compile Include="..\..\..\src\GcodeParserSharp\Models\Gcode.cs" Link="Models\Gcode.cs" />
   </ItemGroup>
 
   <ItemGroup>
